+++ endtoend/evaluators/BUILD.bazel
@@ -4,8 +4,9 @@ go_library(
     name = "go_default_library",
     testonly = True,
     srcs = [
-        "consensus.go",
         "finality.go",
+        "node.go",
+        "types.go",
         "validator.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/endtoend/evaluators",
