+++ endtoend/evaluators/validator.go
@@ -9,13 +9,6 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/params"
 )
 
-// Evaluator defines the structure of the evaluators used to
-// conduct the current beacon state during the E2E.
-type Evaluator struct {
-	Name       string
-	Policy     func(currentEpoch uint64) bool
-	Evaluation func(client eth.BeaconChainClient) error
-}
 
 // ValidatorsAreActive ensures the expected amount of validators are active.
 var ValidatorsAreActive = Evaluator{
